Changes made for dynamic selector (#68)
Changing the GetActiveModel function to public for use in VRExo code for admittance controller in the control loop. With the orientation dynamic being added, the size of different models are different (3D or 6D). Therefore GetActiveModel function was called to retrieve the active using model from DynamicSelector for calculating positions and orientations.

---------

Co-authored-by: Control Computer <FullBodyHapticExo@gmail.com>