Guanyu/feature/spring and lock (#70)
Good news! Now everything works well in simulation. Thank you!

I am not sure why we need to update pointmassfirstorder. it seems that
if we consider it as a low pass filter, the dc gain will contain the
stiffness, and the time constant is a combination of both the damping
and stiffness.

$X = \frac{Force/k}{1+\frac{d}{k}s}$

where k is the stiffness, d is damping, and $\frac{d}{k}$ is the time
constant.

---------

Co-authored-by: lxcleo <70342379+lxcleo@users.noreply.github.com>
Co-authored-by: Guanyu Chen <gychen@seas.upenn.edu>
Co-authored-by: guanyu <gychen98@g.ucla.edu>